MakeVM move to jsonutil
